+++ clean/CROSS.h
@@ -76,11 +76,11 @@ void __namespace__CROSS_keygen(prikey_t *SK,
 /* sign cannot fail */
 void __namespace__CROSS_sign(const prikey_t * SK,
                 const char * m,
-                uint64_t mlen,
+                size_t mlen,
                 CROSS_sig_t * sig);
 
 /* verify returns 1 if signature is ok, 0 otherwise */
 int __namespace__CROSS_verify(const pubkey_t * PK,
                  const char * m,
-                 uint64_t mlen,
+                 size_t mlen,
                  const CROSS_sig_t * sig);
