+++ generate/crypto_sign/cross-rsdpg-128-fast/clean/CROSS.c
@@ -121,7 +121,7 @@ void PQCLEAN_CROSSRSDPG128FAST_CLEAN_CROSS_keygen(prikey_t *SK,
 /* sign cannot fail */
 void PQCLEAN_CROSSRSDPG128FAST_CLEAN_CROSS_sign(const prikey_t *const SK,
         const char *const m,
-        const uint64_t mlen,
+        const size_t mlen,
         CROSS_sig_t *const sig) {
 	/* Wipe any residual information in the sig structure allocated by the
 	 * caller */
@@ -294,7 +294,7 @@ void PQCLEAN_CROSSRSDPG128FAST_CLEAN_CROSS_sign(const prikey_t *const SK,
 /* verify returns 1 if signature is ok, 0 otherwise */
 int PQCLEAN_CROSSRSDPG128FAST_CLEAN_CROSS_verify(const pubkey_t *const PK,
         const char *const m,
-        const uint64_t mlen,
+        const size_t mlen,
         const CROSS_sig_t *const sig) {
 	CSPRNG_STATE_T CSPRNG_state;
